allow setting default scope for generated envoy filters
The default scope is mesh wide, wihch means the EnvoyFilter will be
generated in the root namespace and affect the workloads in all namespaces.
If the environment variable AERAKI_ENABLE_ENVOY_FILTER_NS_SCOPE is set to true,
the EnvoyFilters will be generated in the namespace where the service reside in.

Signed-off-by: huabingzhao <zhaohuabing@gmail.com>